separate experience from projects
