+++ src/App.jsx
@@ -3,7 +3,7 @@ import Hero from "./components/layout/hero/Hero";
 import About from "./components/layout/about/About";
 import Projects from "./components/layout/projects/Projects";
 import Footer from "./components/layout/footer/Footer";
-import DemoCarousel from "./components/slider/slider-test2";
+// import DemoCarousel from "./components/slider/slider-test2";
 
 function App() {
   return (
@@ -11,7 +11,7 @@ function App() {
       <main>
         <Hero />
         <About />
-        <DemoCarousel />
+        {/* <DemoCarousel /> */}
         <Projects />
         <Footer />
       </main>
