+++ src/components/layout/experience/data/experience.js
@@ -1,0 +1,22 @@
+const professionalProjects = [
+  {
+    title: "Groupe Dynamite Inc.",
+    content:
+      "(Dynamite & Garage) Lorem ipsum dolor sit amet consectetur adipisicing elit. Dicta at ad possimus soluta velit voluptatum commodi id. In aspernatur repudiandae, fuga velit obcaecati laboriosam culpa dolorem nobis molestiae itaque dignissimos? ",
+    url: "https://www.dynamiteclothing.com/",
+  },
+  {
+    title: "Cricut",
+    content:
+      "Lorem ipsum dolor sit amet consectetur adipisicing elit. Dicta at ad possimus soluta velit voluptatum commodi id. In aspernatur repudiandae, fuga velit obcaecati laboriosam culpa dolorem nobis molestiae itaque dignissimos? ",
+    url: "https://cricut.com/",
+  },
+  {
+    title: "Cricut Heat Guide",
+    content:
+      "Lorem ipsum dolor sit amet consectetur adipisicing elit. Dicta at ad possimus soluta velit voluptatum commodi id. In aspernatur repudiandae, fuga velit obcaecati laboriosam culpa dolorem nobis molestiae itaque dignissimos? ",
+    url: "https://cricut.com/heatguide",
+  },
+];
+
+export default professionalProjects;
