+++ src/components/layout/projects/Projects.jsx
@@ -1,18 +1,7 @@
 import "./Projects.css";
 import PersonalProjects from "./Personal";
-import ProfessionalProjects from "./Professional";
-import { useState } from "react";
-
-const PROFESSIONAL_TEXT = "Professional";
-const PERSONAL_TEXT = "Personal";
 
 function Projects() {
-  const [projectType, setProjectType] = useState(PROFESSIONAL_TEXT);
-
-  const toggleProject = (type) => {
-    setProjectType(type);
-  };
-
   return (
     <section className="projects-section">
       <div className="projects-text">
@@ -26,20 +15,7 @@ function Projects() {
       </div>
 
       <div className="projects">
-        <div className="project-buttons">
-          <button onClick={() => toggleProject(PROFESSIONAL_TEXT)}>
-            {PROFESSIONAL_TEXT}
-          </button>
-          <button onClick={() => toggleProject(PERSONAL_TEXT)}>
-            {PERSONAL_TEXT}
-          </button>
-        </div>
-
-        {projectType === PROFESSIONAL_TEXT ? (
-          <ProfessionalProjects />
-        ) : (
-          <PersonalProjects />
-        )}
+        <PersonalProjects />
       </div>
     </section>
   );
