+++ src/components/slider/slider-test2.jsx
@@ -1,12 +1,14 @@
 import "react-responsive-carousel/lib/styles/carousel.min.css"; // requires a loader
 import { Carousel } from "react-responsive-carousel";
+import ProfessionalProjects from "../layout/projects/Professional";
+import PersonalProjects from "../layout/projects/Personal";
 
 import MobileImage from "/images/Hero-Image-Mobile.jpg";
 
 function DemoCarousel() {
   return (
     <Carousel>
-      <div>
+      {/* <div>
         <img src={MobileImage} />
         <p className="legend">Legend 1</p>
       </div>
@@ -17,7 +19,9 @@ function DemoCarousel() {
       <div>
         <img src={MobileImage} />
         <p className="legend">Legend 3</p>
-      </div>
+      </div> */}
+      <ProfessionalProjects />
+      <PersonalProjects />
     </Carousel>
   );
 }
